+++ src/TowerModels/MultiTargetsTower.java
@@ -14,7 +14,7 @@ public abstract class MultiTargetsTower extends Tower {
     protected double effectRange;
 
     public MultiTargetsTower(int xPos, int yPos, int level, int initialCost, int upgradeCost,
-            int refundValue, double range, double power, double rateOfFire, double effectRange, Cell cell) {
+            double refundValue, double range, double power, double rateOfFire, double effectRange, Cell cell) {
 
         super(xPos, yPos, level, initialCost, upgradeCost, refundValue, range, power, rateOfFire,
                 true, cell);
