+++ src/TowerModels/SingleTargetTower.java
@@ -10,7 +10,7 @@ import Map.Cell;
 public abstract class SingleTargetTower extends Tower {
 
     public SingleTargetTower(int xPos, int yPos, int level, int initialCost, int upgradeCost,
-            int refundValue, double range, double power, double rateOfFire, Cell cell) {
+            double refundValue, double range, double power, double rateOfFire, Cell cell) {
 
         super(xPos, yPos, level, initialCost, upgradeCost, refundValue, range, power, rateOfFire,
                 false, cell);
